Storage. SQLite
In this tutorial:

- Data storage using SQLite

In the last lesson we looked at the simplest way to store data - the Preferences . But this method is rather limited, and for storing large amounts of structured data is inconvenient. In this lesson we consider the SQLite . This is a database with tables and queries - all in a conventional database.

To begin with, a little theory on the interaction of the application and the database.

In an application to connect to the database, we specify the name of the database and version . At the same time the following situations may arise:

1) DB does not exist . This may be for example in the case of the initial installation program. In this case, the application itself must create the database and all the tables in it. And then it is already working with the newly created database.

2) The database exists , but its version is outdated . This may be the case for program updates. For example the new version of the program needs additional fields in the old tables or new tables. In this case, the application should update existing tables and create new ones if necessary.

3) database exists and its version is relevant . In this case, the application has successfully connected to the database up and running.

As you know, the phrase "the application must" is equivalent to the phrase "the developer should", ie it is our task. To handle the situations described above, we need to create a class , which is the heir to SQLiteOpenHelper . Call it DBHelper. This class will provide us with methods to create or update the database in case of its absence or obsolescence .

the onCreate - a method that is invoked when the database to which we want to connect - not exist

onUpgrade - it will be invoked if we try to connect to the database over the new version than the existing

Let's distribute simple application - reference book of contacts, which will retain the name and an email . Enter the data will be on the screen of the application, and to display information using logs . Usually used for this List (list) - but we have this topic not yet know. Yes, and I do not want to overload the application. The main thing - to learn techniques for working with databases.

Create a project: